+++ CHANGELOG.rst
@@ -31,6 +31,8 @@ Added
     - ``Project.has_role()`` helper ``public`` arg (#1702)
     - Limited guest role alert in ``ProjectDetailView`` (#1705)
     - Search by object UUID (#1546)
+    - ``notify_alert_project`` and ``notify_alert_role`` app settings (#1648)
+    - App alert disabling for project and role updates (#1648)
 
 Changed
 -------